Restructure of components and files, integration of hooks
