+++ src/App.js
@@ -1,0 +1,17 @@
+import React from "react";
+import ReactDOM from "react-dom";
+import Header from "./components/Header";
+import Body from "./components/Body";
+
+const AppLayout = () => {
+  return (
+    <div id="app">
+      <Header />
+      <Body />
+    </div>
+  );
+};
+
+const root = ReactDOM.createRoot(document.getElementById("root"));
+
+root.render(<AppLayout />);
