+++ src/components/Header.js
@@ -1,0 +1,18 @@
+import logo from "../../assets/Images/logo.png";
+
+const Header = () => {
+  return (
+    <div className="header">
+      <div className="logo-container">
+        <img className="logo" src={logo} alt="Logo" />
+      </div>
+      <div className="heading-links">
+        <a href="#">Home</a>
+        <a href="#">About</a>
+        <a href="#">Contact</a>
+      </div>
+    </div>
+  );
+};
+
+export default Header;
