+++ style.css
@@ -34,22 +34,27 @@
     padding: 0px;
     list-style: none;
     margin-block-end: 0px;
-    display: grid;
-    -webkit-box-align: start;
-    align-items: start;
-    grid-template-columns: repeat(4, 1fr);
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-between;
     gap: 32px;
     margin: 32px 16px;
 }
 
 .restaurant-card {
-    display: grid;
-    transition: 0.05s;
+    display: flex;
+    flex-direction: column;
+    width: 18rem;
     cursor: pointer;
     gap: 12px;
-    grid-auto-flow: row;
     justify-content: stretch;
-    -webkit-box-align: center;
-    align-items: center;
-    padding: 0px;
+    padding: 1rem;
+    border: 1px solid rgba(0,0,0,0.2);
+    background-color: whitesmoke;
+}
+
+.restaurant-card img {
+    width: 100%;
+    height: 10rem;
+    object-fit: cover;
 }
